--- client/src/components/organisms/LoginWrapper/style.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Wrapper = styled.div`
-  ${({ theme }) => theme.absoluteCenter};
-  ${({ theme }) => theme.flex.columnCenter};
-  width: 300px;
-`;
