+++ client/src/components/organisms/index.tsx
@@ -1,5 +1,4 @@
 export { default as ProjectCardContainer } from 'components/templates/ProjectCardContainer';
-export { default as LoginWrapper } from 'components/organisms/LoginWrapper';
 export { default as NewProjectCard } from 'components/organisms/NewProjectCard';
 export { default as NodeDetailWrapper } from 'components/organisms/NodeDetailWrapper';
 export { default as ProjectCard } from 'components/organisms/ProjectCard';
