+++ client/src/components/templates/Login/style.ts
@@ -1,10 +1,16 @@
-import styled from "@emotion/styled";
+import styled from '@emotion/styled';
 
 export const Template = styled.div`
   width: 100%;
-  height :100vh;
+  height: 100vh;
   padding: 2rem;
-  color:${(props) => props.theme.color.white};
+  color: ${(props) => props.theme.color.white};
   background-color: ${(props) => props.theme.color.primary1};
   ${(props) => props.theme.flex.columnCenter}
-`
+`;
+
+export const Wrapper = styled.div`
+  ${({ theme }) => theme.absoluteCenter};
+  ${({ theme }) => theme.flex.columnCenter};
+  width: 300px;
+`;
